# Improve performance on large folders by 10x or more