I was kind of frustrated with how long it took to run on large folders when the mail was mostly already downloaded, and I didn't always want to exclude them from my regular backups so I started reading up on IMAP a little bit and found a way to get a really good performance boost.

IMAP allows fetching message IDs for all messages in a folder at once and it's quite a bit faster than making individual requests. Running this on a folder with 4441 messages takes 26s for me now compared to over 8 minutes before this change.

Here's the current master branch processing one folder with 599 messages in just over a minute:
```
[ 4:25pm] ~/Backup/fastmail-email % time imapbackup38.py --ssl --server=imap.fastmail.com --user=someone@example.net --pass=secret --mbox-dir=. --folders=Years/2009
Connecting to 'imap.fastmail.com' TCP port 993, SSL
Logging in as 'someone@example.net'
Finding Folders /(\HasNoChildren) "/" INBOX
(\HasNoChildren \Archive) "/" Archive
...
(\HasChildren) "/" Years
(\HasNoChildren) "/" Years/2004
...
(\HasNoChildren) "/" Years/2021
Finding Folders: 37 folders
File Years.2009.mbox: 600 messages
Folder "Years/2009":
599 messages
New messages: 0
Disconnecting
imapbackup38.py --ssl  --user=someone@example.net  0.89s user 0.17s system 1% cpu 1:04.78 total
```

And this branch on the same folder in 2 seconds:

```
[ 4:27pm] ~/Backup/fastmail-email % time imapbackup38-fast.py --ssl --server=imap.fastmail.com --user=someone@example.net --pass=secret --mbox-dir=. --folders=Years/2009
Connecting to 'imap.fastmail.com' TCP port 993, SSL
Logging in as 'someone@example.net'
Finding Folders /(\HasNoChildren) "/" INBOX
(\HasNoChildren \Archive) "/" Archive
...
(\HasChildren) "/" Years
(\HasNoChildren) "/" Years/2004
...
(\HasNoChildren) "/" Years/2021
Finding Folders: 37 folders
File Years.2009.mbox: 600 messages
Folder "Years/2009":
599 messages
New messages: 0
Disconnecting
imapbackup38-fast.py --ssl --server=imap.fastmail.com      0.75s user 0.10s system 39% cpu 2.137 total
```

Now the current master branch on a folder with 4441 messages:

```
[ 4:27pm] ~/Backup/fastmail-email % time imapbackup38.py --ssl --server=imap.fastmail.com --user=someone@example.net --pass=secret --mbox-dir=. --folders=Years/2021
Connecting to 'imap.fastmail.com' TCP port 993, SSL
Logging in as 'someone@example.net'
Finding Folders /(\HasNoChildren) "/" INBOX
(\HasNoChildren \Archive) "/" Archive
...
(\HasChildren) "/" Years
(\HasNoChildren) "/" Years/2004
...
(\HasNoChildren) "/" Years/2021
Finding Folders: 37 folders
File Years.2021.mbox: 4441 messages
Folder "Years/2021":
4441 messages
New messages: 0
Disconnecting
imapbackup38.py --ssl  --user=someone@example.net  22.47s user 3.51s system 5% cpu 8:13.32 total
```

And this branch on the same folder in 26 seconds:

```
[ 4:36pm] ~/Backup/fastmail-email % time imapbackup38-fast.py --ssl --server=imap.fastmail.com --user=someone@example.net --pass=secret --mbox-dir=. --folders=Years/2021
Connecting to 'imap.fastmail.com' TCP port 993, SSL
Logging in as 'someone@example.net'
Finding Folders /(\HasNoChildren) "/" INBOX
(\HasNoChildren \Archive) "/" Archive
...
(\HasChildren) "/" Years
(\HasNoChildren) "/" Years/2004
...
(\HasNoChildren) "/" Years/2021
Finding Folders: 37 folders
File Years.2021.mbox: 4441 messages
4441 messages
New messages: 0
Disconnecting
imapbackup38-fast.py --ssl --server=imap.fastmail.com      21.53s user 2.91s system 93% cpu 26.233 total
```