@rcarmo 2022-01-02
I’m fine with this, would just like to have someone else test it against more IMAP servers (which is a pain).