@samsonjs 2022-01-02
Yeah that's a good idea. I'm not familiar with IMAP at all and I just found this in a tutorial so I have no idea how portable it is. According to the spec the fetch command accepts a "sequence set" and they use a range in their example, but I can't find a definition of that term anywhere. https://datatracker.ietf.org/doc/html/rfc3501#section-6.4.5

I'm going to do a bit more testing as well by downloading some folders with the current master branch and then this branch and diffing the resulting mbox files. Then updating each set of files and diff again, things like that. I started down that road and found some old, broken emails but I tracked them down and deleted them and can proceed with that testing at some point. But because we're indexing into the resulting data array using num_msgs I'm pretty confident that if something was missing it would blow up. But you never know until you test it.