@samsonjs 2022-01-03
Trying to compare some of my folders is difficult because email is always changing so my inbox, sent, and some other folders don't match. I also had to unify all of the "From " lines because they have generated timestamps in them by default, and sed is unhappy that the files aren't UTF-8. Anyway after crafting up this command to munge all of the From lines, so far it looks like this branch creates identical output to master which is great!

```zsh
for f in *.mbox; do
    echo ">>> $f <<<"
    iconv -f ISO-8859-1 -t UTF8-MAC "$f" | sed -e 's/^From nobody .*$/From nobody Mon Jan  3 11:55:08 2022/' > "$f.new"
done
```

<img width="1176" alt="Screen Shot 2022-01-03 at 3 25 55 PM" src="https://user-images.githubusercontent.com/32487/147991119-aba76201-a763-4815-b5d8-1788c0ab93ad.png">

```zsh
% diff -qruN master-munged fast-munged
Files master-munged/INBOX.mbox and fast-munged/INBOX.mbox differ
Files master-munged/Sent.mbox and fast-munged/Sent.mbox differ
Files master-munged/Spam.mbox and fast-munged/Spam.mbox differ
Files master-munged/Trash.mbox and fast-munged/Trash.mbox differ
Files master-munged/checksums.txt and fast-munged/checksums.txt differ
```