@rcarmo 2022-01-04
Oh, yes, these things are quite hard to test on live mailboxes, for a number of reasons... There were two that gave me some headaches in the past:
* Different IMAP servers have different policies for IMAP message IDs (i.e., message 10 may only be message 10 during a single session, or be directly coupled to the message itself, etc.)
* Deleted messages (flagged for deletion but not expunged or moved to another folder) sometimes lounge about for no apparent reason
Gmail was notorious for handling deletes/moves poorly, for instance.