@rcarmo 2022-01-08
Merging this since I think the approach should work with all IMAP servers during the same backup run.